Fix calculate product price

+++ Jelwelry Sale System At The Store/src/main/java/com/swp391/jewelrysalesystem/controllers/VNPAYController.java
@@ -19,11 +19,11 @@ import java.util.*;
 public class VNPAYController {
 
     @PostMapping("/create_payment")
-    public String createPayment(HttpServletRequest req, @RequestParam("amount") double amount) {
+    public String createPayment(HttpServletRequest req, @RequestParam("amount") int amount) {
         String vnp_Version = "2.1.0";
         String vnp_Command = "pay";
         String orderType = "other";
-        double amountInCents = amount * 100;
+        long amountInCents = amount * 100;
 
         String vnp_TxnRef = VnpayConfig.getRandomNumber(8);
         String vnp_IpAddr = VnpayConfig.getIpAddress(req);
